libentropy: implement a library which abstracts entropy calculation.
Implement libentropy which abstracts entropy calculation and provides
a backend to calculate entropy in Shannon.

User first creates an entropy context (of type struct entropy_ctx) and sets the
->ec_algo field to the algorithm of choice (currently only Shannon entropy
calculation is supported), Then uses the helper function
libentropy_update_ctx() to update the frequency of symbols and other variables
related to entropy calculation, passing the data buffer as an argument. This
function can be called on multiple buffers repeatedly and the context will be
updated as if it's operating on a continuous byte stream. Finally, user calls
libentropy_calculate(), passing the context as an argument. This function
calls the corresponding backend entropy calculation function and sets the
->ec_status field to reflect the result of the operation.

Signed-off-by: gokturk <gokturk@binghamton.edu>