entropy: change the meaning of '-l' option with respect to '-s' option.
Currently '-l' option limits the total size of the file regardless of
how many bytes we skip at the beginning as a result of '-s' option. Change
the file size limit option ('-l') so that processing stops after specified
limit number of bytes are read starting from any offset in the file.

Signed-off-by: gokturk <gokturk@binghamton.edu>