e2ntropy: Check for max device size when iterating over block groups.
If the device is not big enough to fully populate the last block group,
ext4 creates bitmaps for them anyway and marks them as used. This way
allocation never looks at these blocks. However for us, this results
in unnecessary iterations over the bitmap sometimes. Do not look for
blocks beyond the maximum device size.

Signed-off-by: gokturk <gokturk@binghamton.edu>