libe2ntropy: initial commit
libe2ntropy is a library to gather information such entropy, chisq
about the free disk space of ext family file systems.

Right now, it only returns the unused blocks, so that the caller can
use libentropy functions on them directly. We intend to bridge the gap
in the future and perform the calculations inside libe2ntropy
directly.